==> competition/bupt_try/1_process.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

import warnings 

warnings.filterwarnings('ignore')

tmp = pd.read_csv('months1_10/09.csv')
tmp = pd.concat([tmp, pd.read_csv('months1_10/10.csv')])
tmp = pd.concat([tmp, pd.read_csv('months1_10/11.csv')])

train = tmp.groupby(['time_stamp', 'loc_id'], as_index=False)['phone_id'].\
                                                         agg({'num':'count'})

train['month'] = train.time_stamp.apply(lambda x: int(x.split('-')[1]))
train['day'] = train.time_stamp.apply(lambda x: int(x.split('-')[2][:2]))
train['hour'] = train.time_stamp.apply(lambda x: int(x.split('-')[2][-2:]))
train['ymd'] = train.time_stamp.apply(lambda x: x[:-3])
train['dayofweek'] = pd.DatetimeIndex(train['ymd']).dayofweek

train.head(3)

%%time
# 求得星期小时的均值和标准差
train = pd.merge(train, train.groupby(['loc_id', 'dayofweek', 'hour'], as_index=False)['num'].\
                agg({'week_hour_mean':'mean'}), on=['loc_id', 'dayofweek', 'hour'], how='left')
train = pd.merge(train, train.groupby(['loc_id', 'dayofweek', 'hour'], as_index=False)['num'].\
                agg({'week_hour_std':'std'}), on=['loc_id', 'dayofweek', 'hour'], how='left')

%%time
# 取2.4倍的标准差，去掉大概0.5%的异常值数据
train['del'] = train.apply(lambda x: 0 if (x['num'] <= x['week_hour_mean']+2.4*x['week_hour_std']) and \
                           (x['num'] >= x['week_hour_mean']-2.4*x['week_hour_std']) else 1,\
                            axis=1)

# 取9月4号开学后的数据，去掉十一假期
train = train.loc[(train['ymd']>='2017-09-04')&\
                  ~((train['ymd']>='2017-10-01')&(train['ymd']<='2017-10-08'))]

train = train.loc[train['del']!=1]

del train['del']

train.to_csv('data/train_9_11.csv', index=None)
