==> competition/bupt_try/2_model.ipynb <==
import pandas as pd
import numpy as np
from sklearn.metrics import mean_squared_error
from scipy.optimize import minimize
import lightgbm as lgb
from sklearn.ensemble import RandomForestRegressor
from sklearn.ensemble import GradientBoostingRegressor

import warnings
warnings.filterwarnings('ignore')

train = pd.read_csv('data/train_9_11.csv')
weather = pd.read_csv('data/weather_3_12.csv')

len(train),len(test)

# 是否工作日，周末
train['isweekend'] = train['dayofweek'].apply(lambda x: 1 if x==5 or x==6 else 0)
train['isweekday'] = 1 - train['isweekend']
train['day'] = train['ymd'].apply(lambda x: int(x[-2:]))
train['month'] = train['ymd'].apply(lambda x: int(x[5:7]))

train = pd.merge(train, weather, on=['ymd'], how='left')

# 地点小时均值，带衰减权重
def getmean(df):
    num = df['num'].values
    weights = [ 0.85**i for i in range(len(num),0,-1) ] 
    return sum(num*weights)/sum(weights)
train = pd.merge(train,\
train.loc[(train['ymd']<'2017-11-24')&(train.month.isin([10,11])),:].groupby(['loc_id', 'hour']).apply(getmean).reset_index().rename(columns={0:'loc_hour_mean'}),\
                on=['loc_id', 'hour'], how='left')

# 地点星期小时均值，带衰减权重
train = pd.merge(train,\
train.loc[(train['ymd']<'2017-11-24')&(train.month.isin([10,11])),:].groupby(['loc_id', 'dayofweek', 'hour']).apply(getmean).reset_index().rename(columns={0:'loc_week_hour_mean'}),\
                on=['loc_id', 'dayofweek', 'hour'], how='left')

# 小时分区
hour_bins = [-1, 6,19, 23]
train['hour_cut'] = pd.cut(train.hour, hour_bins, labels=False)

# 月份分区，上、中、下旬
day_bins = [0, 10, 20, 31]
train['day_cut'] = pd.cut(train.day, day_bins, labels=False)

# 地点简单分类，食堂，教学楼，公寓
train['loc_type'] = 0
train.loc[train['loc_id'].isin([8,10,12,29]),'loc_type'] = 1  # 食堂
train.loc[train['loc_id'].isin([16,33,27,15,22,24,14,20]),'loc_type'] = 2 # 教室

features = [
            'loc_id','loc_type',
            'month','day','day_cut',
            'hour','hour_cut',
            'dayofweek','isweekday', 'isweekend',
            'loc_hour_mean','loc_week_hour_mean',
            'maxTem','minTem',
            'wea','pm','rain'
             ]
label = ['num']

# 仅用11月数据训练模型
X_train = train.loc[(train['ymd']<'2017-11-24')&(train.month.isin([11])), features]
X_valid = train.loc[(train['ymd']>'2017-11-23'), features]
y_train = train.loc[(train['ymd']<'2017-11-24')&(train.month.isin([11])), label].num.values
y_valid = train.loc[(train['ymd']>'2017-11-23'), label].num.values
print(X_train.shape)
print(X_valid.shape)

lgb_params = {
        'boosting': 'gbrt',
        'max_depth': 10,
        'num_leaves': 60,
        'learning_rate': 0.01,
        'colsample_bytree': 0.5,
        'objective': 'regression',
        'metric': 'rmse',
        'min_data_in_leaf':10,
        'verbose': -1,
        'nthread': 12,
    }

d_train = lgb.Dataset(X_train, label=y_train)
d_valid = lgb.Dataset(X_valid, label=y_valid)
watchlist = [d_train, d_valid]

model = lgb.train(lgb_params, train_set=d_train, num_boost_round=2000, valid_sets=watchlist, \
early_stopping_rounds=20, verbose_eval=100)
lgb_pred = model.predict(X_valid)

## blend

def rmse(y, y_pred):
    assert len(y) == len(y_pred)
    return (sum((y_pred - y)** 2.0)/len(y)) ** 0.5

%%time
rf = RandomForestRegressor(n_estimators=200, max_depth=8, max_features='sqrt', n_jobs=16)
rf.fit(X_train, y_train)
rf_pred = rf.predict(X_valid)
print('rf rmse:', rmse(y_valid, rf_pred))

%%time
gbm = GradientBoostingRegressor(learning_rate=0.01,
                                n_estimators=1000,
                                max_depth=6,
                                max_features='sqrt',
                                criterion='mse',
                               )
gbm.fit(X_train, y_train)
gbm_pred = gbm.predict(X_valid)
print('gbm rmse:', rmse(y_valid, gbm_pred))

rule_ans = train.loc[(train['ymd']<'2017-11-24')&(train.month.isin([10,11]))].groupby(['loc_id', 'dayofweek', 'hour']).\
    apply(getmean).reset_index().rename(columns={0:'num_1'})

rule_pred = pd.merge(X_valid, rule_ans, on=['loc_id','dayofweek','hour'], how='left').fillna(0)['num_1'].values

rmse(y_valid, rule_pred)

%%time
# 使用验证集的预测结果来确定模型融合的参数
blend_train= []
res_list = []
weights_list = []
blend_train.append(lgb_pred)
blend_train.append(rf_pred)
blend_train.append(gbm_pred)
blend_train.append(rule_pred)
blend_train = np.array(blend_train)

def rmse_min_func(weights):
    final_prediction = 0
    for weight, prediction in zip(weights, blend_train):
        final_prediction += weight * prediction
    return np.sqrt(mean_squared_error(y_valid, final_prediction))

for k in range(10):
    starting_values = np.random.uniform(size=len(blend_train))
    bounds = [(-1, 1)] * len(blend_train)

    res = minimize(rmse_min_func,
                   starting_values,
                   method='L-BFGS-B',
                   bounds=bounds,
                   options={'disp': False,
                            'maxiter': 100000})
    res_list.append(res['fun'])
    weights_list.append(res['x'])
    print('{iter}\tScore: {score}\tWeights: {weights}'.format(
        iter=(k + 1),
        score=res['fun'],
        weights='\t'.join([str(item) for item in res['x']])))

## submit

# 重新训练一次模型，使用11月全部数据
d_train = lgb.Dataset(train.loc[train.month.isin([11]), features], label=train.loc[train.month.isin([11])].num.values)
watchlist = [d_train]
model = lgb.train(lgb_params, train_set=d_train, num_boost_round=300, valid_sets=watchlist, \
early_stopping_rounds=None, verbose_eval=100)

rf.fit(train.loc[train.month.isin([11]), features], \
       train.loc[train.month.isin([11])].num.values)
gbm.fit(train.loc[train.month.isin([11]), features], \
        train.loc[train.month.isin([11])].num.values)

test = pd.read_csv('data/test_data_12.csv')

test['hour'] = test.time_stamp.apply(lambda x: int(x[-2:]))
test['month'] = test.time_stamp.apply(lambda x: int(x.split('-')[1]))
test['day'] = test.time_stamp.apply(lambda x: int(x.split('-')[2][:2]))

test['dayofweek'] = pd.DatetimeIndex(test.time_stamp).dayofweek
test['isweekend'] = test['dayofweek'].apply(lambda x: 1 if x==5 or x==6 else 0)
test['isweekday'] = 1 - test['isweekend']
test['ymd'] = test.time_stamp.apply(lambda x: x[:10])
test = pd.merge(test, weather, on='ymd', how='left')

test = pd.merge(test,\
                train.loc[train.month.isin([10,11])].groupby(['loc_id', 'hour']).apply(getmean).reset_index().rename(columns={0:'loc_hour_mean'}),\
                on=['loc_id', 'hour'], how='left')

test = pd.merge(test,\
                train.loc[train.month.isin([10,11])].groupby(['loc_id', 'dayofweek', 'hour']).apply(getmean).reset_index().rename(columns={0:'loc_week_hour_mean'}),\
                on=['loc_id', 'dayofweek', 'hour'], how='left')

hour_bins = [-1, 6, 19, 23]
test['hour_cut'] = pd.cut(test.hour, hour_bins, labels=False)
day_bins = [0, 10, 20, 31]
test['day_cut'] = pd.cut(test.day, day_bins, labels=False)

test['loc_type'] = 0
test.loc[test['loc_id'].isin([8,10,12,29]),'loc_type'] = 1
test.loc[test['loc_id'].isin([16,33,27,15,22,24,14,20]),'loc_type'] = 2

rule_ans_test = train.loc[train.month.isin([10,11])].groupby(['loc_id', 'dayofweek', 'hour']).\
apply(getmean).reset_index().rename(columns={0:'num_1'})

rule_res = \
pd.merge(test, rule_ans_test, on=['loc_id','dayofweek','hour'], how='left')['num_1'].values

%%time
test['num_of_people'] = \
weights_list[-1][0] * model.predict(test[features]) +\
weights_list[-1][1] * rf.predict(test[features]) +\
weights_list[-1][2] * gbm.predict(test[features]) +\
weights_list[-1][3] * rule_res

test[['loc_id', 'time_stamp', 'num_of_people']].to_csv('submit/submit_gbm_normal_lgb.csv', index=False)
